add scoring + lives

+++ src/main/java/de/cargame/controller/PlayerController.java
@@ -1,7 +1,7 @@
 package de.cargame.controller;
 
 import de.cargame.api.PlayerApi;
-import de.cargame.model.entity.Player;
+import de.cargame.model.entity.player.Player;
 import de.cargame.model.handler.PlayerHandler;
 
 public class PlayerController implements PlayerApi {
@@ -19,7 +19,7 @@ public class PlayerController implements PlayerApi {
 
     @Override
     public int getScore(String playerId) {
-        return playerHandler.getScore(playerId);
+        return (int) playerHandler.getScore(playerId);
     }
